* add repositories implementations

+++ app/src/main/java/com/kozak/pw/domain/repository/NewsItemRepository.kt
@@ -5,5 +5,5 @@ import com.kozak.pw.domain.news.NewsItem
 interface NewsItemRepository {
     fun addNews(newsItem: NewsItem)
     fun getNewsList(): List<NewsItem>
-    fun readNews(newsItemId: Int)
+    fun readNews(newsItemId: Long)
 }
